Move the list collage card into the Android design system
